+++ DAW/BAE/Unidad-2/Tarea-6/src/registro_de_empleados.sql
@@ -16,6 +16,7 @@ USE registro_de_empleados_db;
 --  ═╩╝┴└─└─┘┴     ╩ ┴ ┴└─┘┴─┘└─┘
 
 -- Eliminar las tablas si ya existen (para evitar errores al crear las tablas).
+DROP TABLE IF EXISTS Departamentos;
 DROP TABLE IF EXISTS Empleados;
 DROP TABLE IF EXISTS Telefonos;
 
@@ -23,11 +24,19 @@ DROP TABLE IF EXISTS Telefonos;
 --  ║  ├┬┘├┤ ├─┤ │ ├┤    ║ ├─┤├┴┐│  ├┤ 
 --  ╚═╝┴└─└─┘┴ ┴ ┴ └─┘   ╩ ┴ ┴└─┘┴─┘└─┘
 
+-- Crear tabla "Departamentos".
+CREATE TABLE Departamentos (
+    id_departamento INT AUTO_INCREMENT PRIMARY KEY,
+    nombre_departamento VARCHAR(50) NOT NULL,
+    CONSTRAINT Unico_Departamento UNIQUE (nombre_departamento)
+);
+
 -- Crear tabla "Empleados".
 CREATE TABLE Empleados (
     id_empleado INT AUTO_INCREMENT PRIMARY KEY,
-    nombre VARCHAR(100) NOT NULL,
-    departamento VARCHAR(50) NOT NULL
+    id_departamento INT,
+    nombre_empleado VARCHAR(100) NOT NULL,
+    FOREIGN KEY (id_departamento) REFERENCES Departamentos(id_departamento)
 );
 
 -- Crear tabla "Telefonos".
@@ -41,11 +50,17 @@ CREATE TABLE Telefonos (
 --  ║│││└─┐├┤ ├┬┘ │   ╚╗╔╝├─┤│  │ │├┤ └─┐
 --  ╩┘└┘└─┘└─┘┴└─ ┴    ╚╝ ┴ ┴┴─┘└─┘└─┘└─┘
 
--- Insertar en la tabla "Empleados".
-INSERT INTO Empleados (id_empleado, nombre, departamento)
+-- Insertar en la tabla "Departamentos".
+INSERT INTO Departamentos (id_departamento, nombre_departamento)
 VALUES
-    (1, 'Carlos R.', 'Ventas'),
-    (2, 'Laura M.', 'Finanzas');
+    (1, 'Ventas'),
+    (2, 'Finanzas');
+
+-- Insertar en la tabla "Empleados".
+INSERT INTO Empleados (id_empleado, id_departamento, nombre_empleado)
+VALUES
+    (1, 1, 'Carlos R.'),
+    (2, 2, 'Laura M.');
 
 -- Insertar en la tabla "Telefonos".
 INSERT INTO Telefonos (telefono, id_empleado)