+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: cb233ca - Update DAW/BAE/Unidad-2/Tarea-6
+| Author: fjrodafo - Tue, 18 Feb 2025 14:40:36 +0000
+| 
 * Commit: 0f04e4c - Update DAW/BAE/Unidad-2/Tarea-6
 | Author: fjrodafo - Tue, 18 Feb 2025 13:55:25 +0000
 |